add build tags
